# api/openapi.yaml
openapi: 3.1.0
info:
  title: Curio Daemon API
  version: 0.1.0
  description: |
    HTTP+JSON API exposed by `curio-daemon` on `127.0.0.1:8765` (the port is
    configurable; the host must be loopback). Consumed by the `curio` CLI,
    the `curio-mcp` sidecar, and any future client.

    This file is the contract. Tests in `internal/api` check it against the
    router (every route is documented, every documented operation is
    routed) and against live responses (every operation's real answers
    validate against their schemas, with undeclared fields treated as
    errors). The clients in `internal/client` are written by hand; code
    generation is deferred.

    Conventions:
      - IDs are UUIDs. Timestamps are RFC 3339, UTC.
      - Unset optional fields are omitted; the API never sends `null`.
      - Errors follow RFC 7807 (`application/problem+json`). Every response
        carries an `X-Request-Id` header, and every problem repeats it as
        `request_id`, which is also on the daemon's log lines for the
        request; `instance` is the request path. A 404 names what is
        missing. A 500's detail is the raw error text.
      - Responses are read tolerantly: they may gain fields and enum values
        within /v1, and clients ignore what they don't know. Requests are
        strict: an unknown field is a 400 that names it and, since it
        usually means a newer client talking to an older daemon, says to
...
        again.
...

...

...
    this API as JSON.
servers:
...
paths:
...
  /v1/stats:
    get:
      tags: [system]
      operationId: getStats
      summary: Corpus and queue counts
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Stats" }
        default: { $ref: "#/components/responses/Problem" }
...
  /v1/bookmarks:
    get:
      tags: [bookmarks]
      operationId: listBookmarks
      summary: List bookmarks, newest first
      parameters:
        - in: query
          name: source
          schema: { $ref: "#/components/schemas/Source" }
        - in: query
          name: folder
          description: >
            The folder and every folder under it, matched on path segments
            and case-sensitively; a trailing "/" is ignored and "/" alone is
            no filter.
          schema: { type: string }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: One page of bookmarks, each with its document's state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BookmarkList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        default: { $ref: "#/components/responses/Problem" }

    post:
      tags: [bookmarks]
      operationId: createBookmark
      summary: Add one bookmark; a URL new to the corpus gets a fetch job
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateBookmarkRequest" }
      responses:
        "201":
          description: Bookmark created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BookmarkCreated" }
        "400":
          description: url is missing, malformed, not http(s), or has no host; or the body is malformed
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "409":
          description: A manual bookmark for this normalized URL already exists
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/bookmarks/import:
    post:
      tags: [bookmarks]
      operationId: importBookmarks
      summary: Import a batch of parsed bookmarks
      description: |
        Synchronous: the bookmarks are saved and counted before the
        response, one transaction each, and a fetch job is enqueued for
        every URL new to the corpus; the fetches run in the background. The
        client parses the browser's bookmark file and sends the result, in
        batches (the CLI sends 500 at a time). A bookmark that already
        exists for this source is skipped; one curio can't fetch is
        filtered and counted under its reason.

        With dry_run, nothing is written: the response is the one this
        body would get against the library as it is now, with dry_run
        true and new_urls, the URLs the import would fetch (so
        jobs_enqueued counts the fetches it would enqueue). `curio up`
        counts a source's new bookmarks this way, sending its distinct
        URLs alone, up to 10,000 a request.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ImportRequest" }
      responses:
        "200":
          description: What the batch did
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportResult" }
        "400":
          description: Unknown source, an empty batch, or a malformed body
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "413":
          description: The body is over 32 MiB
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/bookmarks/{id}:
    parameters:
      - { $ref: "#/components/parameters/ID" }
    get:
      tags: [bookmarks]
      operationId: getBookmark
      summary: Get one bookmark
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Bookmark" }
        "404": { $ref: "#/components/responses/NotFound" }
        default: { $ref: "#/components/responses/Problem" }
    delete:
      tags: [bookmarks]
      operationId: deleteBookmark
      summary: Delete a bookmark
      description: Removes the bookmark only; its document stays in the corpus.
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
        default: { $ref: "#/components/responses/Problem" }

  # ---------- documents ----------
  /v1/documents:
    get:
      tags: [documents]
      operationId: listDocuments
      summary: List documents, most recently updated first
      description: >
        Every filter given must match. Each is checked on the documents the
        list walks in its order, so a filter that matches few documents
        reads more of the list to fill a page.
      parameters:
        - in: query
          name: state
          schema: { $ref: "#/components/schemas/DocState" }
        - in: query
          name: content_type
          schema: { $ref: "#/components/schemas/ContentType" }
        - in: query
          name: host
          description: >
            Documents whose http or https URL has exactly this host, matched
            case-insensitively, every character literal (no wildcards, no
            subdomains), as the search host filter matches it.
          schema: { type: string }
        - in: query
          name: folder
          description: >
            Documents with a bookmark in this folder or any folder under it,
            matched as the bookmark list's folder filter: on path segments
            and case-sensitively; a trailing "/" is ignored and "/" alone is
            no filter.
          schema: { type: string }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: One page of documents, each with its last error and markdown path
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DocumentList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/documents/{id}:
    parameters:
      - { $ref: "#/components/parameters/ID" }
    get:
      tags: [documents]
      operationId: getDocument
      summary: Get document metadata and its current extraction
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
        "404": { $ref: "#/components/responses/NotFound" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/documents/{id}/content:
    parameters:
      - { $ref: "#/components/parameters/ID" }
    get:
      tags: [documents]
      operationId: getDocumentContent
      summary: Get the extracted markdown
      responses:
        "200":
          description: OK
          content:
            text/markdown:
              schema: { type: string }
        "404":
          description: >
            No such document, or it has no content yet (never fetched, or
            its markdown was deleted from disk).
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/documents/{id}/related:
    parameters:
      - { $ref: "#/components/parameters/ID" }
      - in: query
        name: k
        description: Max related documents. Out of range or malformed means the default.
        schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
    get:
      tags: [documents]
      operationId: findRelatedDocuments
      summary: Find documents related to this one by embedding similarity
      description: >
        Ranks other documents by vector similarity to this document's stored
        chunk embeddings (mean-pooled; no query text involved, so Ollama is
        not needed). The source document is excluded, and so are failed and
        dead documents, whose chunks come from an earlier fetch. Returns an
        empty items list when the document has no indexed chunks yet. Scores
        are raw vector similarities (1/(1+L2 distance), 0..1) and are NOT
        comparable with /v1/search's RRF-fused scores.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RelatedResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/documents/{id}/refetch:
    parameters:
      - { $ref: "#/components/parameters/ID" }
      - in: query
        name: force
        description: 1 or true (any case) refetches a dead document too.
        schema: { type: string }
    post:
      tags: [documents]
      operationId: refetchDocument
      summary: Reset the document to pending and enqueue a fresh fetch job
      description: >
        The state reset and the new job commit together: on failure the
        document keeps its state and no job exists. The current extraction is
        kept until the new fetch replaces it.
      responses:
        "202":
          description: Fetch job enqueued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobRef" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: The document is dead (a confirmed dead link) and force was not given
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/documents/refetch-all:
    parameters:
      - in: query
        name: state
        description: "Only refetch documents in this state. Default: pending, fetched and failed; dead documents only when asked for."
        schema: { $ref: "#/components/schemas/DocState" }
    post:
      tags: [documents]
      operationId: refetchAllDocuments
      summary: Reset matching documents to pending and enqueue a fetch job for each
      description: >
        All-or-nothing: every matching document is reset and gets a fetch job
        in one transaction, or nothing changes. There is no parent job to
        poll; watch GET /v1/jobs or GET /v1/stats instead.
      responses:
        "202":
          description: Fetch jobs enqueued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobsEnqueued" }
        "400": { $ref: "#/components/responses/BadRequest" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/documents/{id}/reindex:
    parameters:
      - { $ref: "#/components/parameters/ID" }
    post:
      tags: [documents]
      operationId: reindexDocument
      summary: Enqueue an index job over the document's current extraction
      description: >
        Re-chunks and re-embeds the markdown already on disk; nothing is
        refetched and the document's state is left alone.
      responses:
        "202":
          description: Index job enqueued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobRef" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: The document has no extraction to reindex; refetch it first
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/documents/reindex-all:
    parameters:
      - in: query
        name: state
        description: "Reindex the documents in this state that have content. Default: fetched."
        schema: { $ref: "#/components/schemas/DocState" }
    post:
      tags: [documents]
      operationId: reindexAllDocuments
      summary: Enqueue an index job for every document in a state that has content
      description: >
        Index jobs change no document state, so a partial run strands
        nothing; a failure stops the run and the 500 says how many jobs were
        enqueued before it. There is no parent job to poll. With
        state=fetched, the default, the fetched documents are every
        searchable one, so once all their jobs are enqueued the embedding
        drift baseline is reset: the recorded model digest and Ollama
        version are cleared, and the daemon's next check records the build
        serving now, which clears embedding_drift on healthz. A failed
        reset is a 500 whose detail says all the jobs were enqueued. Other
        states leave the baseline alone.
      responses:
        "202":
          description: Index jobs enqueued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobsEnqueued" }
        "400": { $ref: "#/components/responses/BadRequest" }
        default: { $ref: "#/components/responses/Problem" }

  # ---------- search ----------
  /v1/search:
    post:
      tags: [search]
      operationId: search
      summary: Hybrid BM25 + vector search
      description: >
        Runs BM25 and vector retrieval concurrently and fuses them with RRF.
        Failed and dead documents are not searched: they may keep the chunks
        of an earlier fetch, which no longer describe what the URL serves. A
        pending document (a refetch in flight) is. If the embedding model is
        unavailable (Ollama down, or slower than search.embed_timeout_seconds),
        the keyword results are still returned with degraded=true and a
        warning explaining why; a keyword-index failure is a 500.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SearchRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SearchResponse" }
        "400":
          description: Missing query, k outside 0..100, an unknown field, or a malformed body
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default: { $ref: "#/components/responses/Problem" }

  # ---------- insight ----------
  /v1/interests:
    get:
      tags: [insight]
      operationId: listInterests
      summary: List inferred interests (labeled topic clusters)
      description: >
        Returns the interests from the latest completed clustering run, largest
        first, with run metadata. An interest is a labeled cluster. Returns an
        empty list when clustering has not run yet.
      parameters:
        - in: query
          name: limit
          description: Max interests. Out of range or malformed means the default.
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
        - in: query
          name: members
          description: Member documents per interest; 0 for none. Out of range or malformed means the default.
          schema: { type: integer, minimum: 0, maximum: 100, default: 5 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InterestList" }
        default: { $ref: "#/components/responses/Problem" }
  /v1/interests/rebuild:
    post:
      tags: [insight]
      operationId: rebuildInterests
      summary: Recompute interests from the current corpus
      description: Enqueues a clustering job. Poll the returned job id via /v1/jobs/{id}.
      responses:
        "202":
          description: Clustering job enqueued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobRef" }
        "409":
          description: Insight layer is disabled (config insight.enabled = false)
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  parameters:
...
    Limit:
      in: query
      name: limit
      description: Rows per page. Out of range or malformed means the default.
      schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
    Cursor:
      in: query
      name: cursor
      description: The next_cursor of the previous page.
      schema: { type: string }
...
  schemas:
...
    DocState:
      type: string
      enum: [pending, fetched, failed, dead]
...
    Source:
      type: string
      enum: [chrome, safari, firefox, html, manual]

    # ---------- system ----------
...